Move request dispatcher to a separate file
